requirements.txt
streamlit==1.41.1
google-generativeai==0.8.5
gspread==6.1.4
google-auth==2.38.0
google-auth-oauthlib==1.2.1
pandas==2.2.3
python-dateutil==2.9.0
